pyproject.toml
[tool.poetry]
name = "serial-job-applier"
version = "0.1.0"
description = "LinkedIn Job Application Automation System"
authors = ["Joan Canossa <jfcpersonal99@gmail.com>"]
readme = "README.md"
packages = [
    {include = "src"},
]

[tool.poetry.dependencies]
python = "^3.12,<3.13"
langgraph = "^0.6.7"
langchain = "^0.3.27"
langchain-community = "^0.3.29"
fastmcp = "^2.12.3"
mcp = "^1.0.0"
selenium = "^4.15.0"
beautifulsoup4 = "^4.12.0"
requests = "^2.31.0"
webdriver-manager = "^4.0.0"
undetected-chromedriver = "^3.5.0"
lxml = "^4.9.0"
fake-useragent = "^1.4.0"
loguru = "^0.7.0"
pypdf = "^5.0.0"
pdfplumber = "^0.10.0"
langchain-huggingface = "^0.3.1"
click = "^8.1.0"
rich = "^13.7.0"
pydantic = "^2.0.0"
pyyaml = "^6.0.0"
typer = ">=0.15.0"
python-dotenv = "^1.0.0"
langfuse = "^2.0.0"
setuptools = "^80.9.0"
uvicorn = {extras = ["standard"], version = "^0.37.0"}
httpx = "^0.28.1"
websockets = ">=14.0"
pandas = "^2.2.0"
langchain-openai = "^0.3.0"
fastapi = "^0.115.0"
confluent-kafka = "^2.6.0"
pytest = "^9.0.2"
basedpyright = "^1.37.2"
sqlalchemy = "^2.0.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
black = "^25.9.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
